+++ packages/aligner/spec/aligner-spec.js
@@ -19,6 +19,10 @@ describe("Aligner", () =>  {
     });
 
     waitsForPromise(() => {
+      return atom.packages.activatePackage('aligner-coffeescript');
+    });
+
+    waitsForPromise(() => {
       return atom.workspace.open('aligner-sample.coffee');
     });
 
@@ -26,10 +30,6 @@ describe("Aligner", () =>  {
       editor = atom.workspace.getActiveTextEditor();
       editorView = atom.views.getView(editor);
     });
-  });
-
-  afterEach(() => {
-    atom.config.unset('aligner');
   });
 
   it("should align two lines correctly", () => {
@@ -46,7 +46,7 @@ describe("Aligner", () =>  {
 
   it("should ailgn correctly with config update", () => {
     editor.setCursorBufferPosition([6, 1]);
-    atom.config.set('aligner.:-alignment', 'left');
+    atom.config.set('aligner-coffeescript.:-alignment', 'left');
     atom.commands.dispatch(editorView, 'aligner:align');
     expect(editor.lineTextForBufferRow(6)).toBe("  foo       : bar");
   });
@@ -180,6 +180,10 @@ describe("Aligning javascript", () => {
 
     waitsForPromise(() => {
       return atom.packages.activatePackage('language-javascript');
+    });
+
+    waitsForPromise(() => {
+      return atom.packages.activatePackage('aligner-javascript');
     });
 
     waitsForPromise(() => {