+++ packages/aligner/spec/operator-config-spec.js
@@ -1,8 +1,7 @@
 'use strict';
 
 const operatorConfig = require('../lib/operator-config');
-const config = require('../config');
-const extend = require('extend');
+const path = require('path');
 const Disposable = require('atom').Disposable;
 
 const cssProvider = {
@@ -44,31 +43,79 @@ const cssProvider = {
 };
 
 describe('Operator Config', () => {
+  let editor
+
   beforeEach(() => {
+    atom.project.setPaths([path.join(__dirname, 'fixtures')]);
+
     operatorConfig.removeAll();
-    operatorConfig.add('aligner', config);
+
+    waitsForPromise(() => {
+      return atom.packages.activatePackage('language-javascript');
+    });
+
+    waitsForPromise(() => {
+      return atom.packages.activatePackage('aligner');
+    });
+
+    waitsForPromise(() => {
+      return atom.packages.activatePackage('aligner-javascript');
+    });
+
+    waitsForPromise(() => {
+      return atom.workspace.open('helper-sample.js')
+      .then((o) => {
+        editor = o;
+      });
+    });
   });
 
   describe('getConfig', () => {
     it('should get the config from config.json', () => {
-      expect(operatorConfig.getConfig('=')).toBeDefined();
+      expect(operatorConfig.getConfig('=', editor)).toBeDefined();
     });
 
     it('should return null when character is not supported', () => {
-      expect(operatorConfig.getConfig('-')).toBeUndefined();
+      expect(operatorConfig.getConfig('-', editor)).toBeUndefined();
     });
 
     it('should be able to get prefixed operator config', () => {
-      expect(operatorConfig.getConfig('+=')).toBeDefined();
+      expect(operatorConfig.getConfig('+=', editor)).toBeDefined();
     });
 
     it('should get the right provider', () => {
-      operatorConfig.add('aligner-css', cssProvider);
-      expect(operatorConfig.getConfig(':', '.source.css')).toBeDefined();
+      let cssEditor
+
+      waitsForPromise(() => {
+        return atom.packages.activatePackage('language-css');
+      });
+
+      waitsForPromise(() => {
+        return atom.workspace.open('test.css')
+        .then((o) => {
+          cssEditor = o;
+        });
+      });
+
+      runs(() => {
+        operatorConfig.add('aligner-css', cssProvider);
+        expect(operatorConfig.getConfig(':', cssEditor)).toBeDefined();
+      });
     });
 
     it('should not be able to get any config', () => {
-      expect(operatorConfig.getConfig('=>', '.source.ruby')).toBeUndefined();
+      let rubyEditor
+
+      waitsForPromise(() => {
+        return atom.workspace.open('test.rb')
+        .then((o) => {
+          rubyEditor = o;
+        });
+      });
+
+      runs(() => {
+        expect(operatorConfig.getConfig('=>', rubyEditor)).toBeUndefined();
+      });
     });
   });
 
@@ -108,8 +155,9 @@ describe('Operator Config', () => {
 
   describe('canAlignWith', () => {
     let characterConfig = null;
+
     beforeEach(() => {
-      characterConfig = operatorConfig.getConfig('=');
+      characterConfig = operatorConfig.getConfig('=', editor);
     });
 
     it('should return true if they are the same', () => {
@@ -127,12 +175,12 @@ describe('Operator Config', () => {
 
   describe('isPrefixed', () => {
     it('should return true when operator has prefix', () => {
-      let characterConfig = operatorConfig.getConfig('+=');
+      let characterConfig = operatorConfig.getConfig('+=', editor);
       expect(operatorConfig.isPrefixed('+=', characterConfig)).toBe(true);
     });
 
     it('should return false when operator does not have prefix', () => {
-      let characterConfig = operatorConfig.getConfig('=');
+      let characterConfig = operatorConfig.getConfig('=', editor);
       expect(operatorConfig.isPrefixed('=', characterConfig)).toBe(false);
     });
   });
@@ -144,9 +192,9 @@ describe('Operator Config', () => {
           alignment: 'right'
         }
       };
-      operatorConfig.updateSetting('aligner', setting);
+      operatorConfig.updateSetting('aligner-javascript', setting);
 
-      expect(operatorConfig.getConfig('+=').alignment).toBe('right');
+      expect(operatorConfig.getConfig('+=', editor).alignment).toBe('right');
     });
   });
 
@@ -167,33 +215,14 @@ describe('Operator Config', () => {
     });
   });
 
-  describe('getAtomConfig', () => {
-    let atomConfig = null;
-    beforeEach(() => {
-      atomConfig = operatorConfig.getAtomConfig();
-    });
-
-    it('should create key-value pairs for Atom config', () => {
-      expect(atomConfig['=-alignment']).toBeDefined();
-    });
-
-    it('should not create key-value pairs for prefixed operators', () => {
-      expect(atomConfig['+=-alignment']).toBeUndefined();
-    });
-
-    it('should contain title', () => {
-      expect(atomConfig['=-alignment'].title).toBeDefined();
-    });
-  });
-
   describe('updateConfigWithAtom', () => {
     it('should update with Atom setting changes', () => {
       let setting = {
         '=-alignment': 'right'
       };
-      operatorConfig.updateConfigWithAtom('aligner', setting);
+      operatorConfig.updateConfigWithAtom('aligner-javascript', setting);
 
-      expect(operatorConfig.getConfig('+=').alignment).toBe('right');
+      expect(operatorConfig.getConfig('+=', editor).alignment).toBe('right');
     });
   });